+++ ogc-example/src/main/scala/geotrellis/server/ogc/conf/package.scala
@@ -18,10 +18,9 @@ package geotrellis.server.ogc
 
 import geotrellis.server.ogc.wms.wmsScope
 import geotrellis.server.ogc.style._
-
 import geotrellis.proj4.CRS
 import geotrellis.vector.Extent
-import geotrellis.raster.TileLayout
+import geotrellis.raster.{ResampleMethod, TileLayout, resample}
 import geotrellis.raster.render.{ColorMap, ColorRamp}
 
 import com.azavea.maml.ast._
@@ -151,6 +150,21 @@ package object conf {
         case Some(layout) => layout
         case None => throw new Exception(s"Invalid layout: $layout. Should be (layoutCols, layoutRows, tileCols, tileRows)")
       }
+    }
+
+  implicit val resampleMethodReader: ConfigReader[ResampleMethod] =
+    ConfigReader[String].map {
+      case "nearest-neighbor"  => resample.NearestNeighbor
+      case "bilinear"          => resample.Bilinear
+      case "cubic-convolution" => resample.CubicConvolution
+      case "cubic-spline"      => resample.CubicSpline
+      case "lanczos"           => resample.Lanczos
+      case "average"           => resample.Average
+      case "mode"              => resample.Mode
+      case "median"            => resample.Median
+      case "max"               => resample.Max
+      case "min"               => resample.Min
+      case "sum"               => resample.Sum
     }
 
   /** An alternative AST reading strategy that uses a separate json file */