+++ ogc/src/main/scala/geotrellis/server/ogc/wcs/WcsModel.scala
@@ -30,11 +30,11 @@ case class WcsModel(
     val filteredSources = sources.find(p.toQuery)
     logger.debug(s"Filtering sources: ${sources.store.length} -> ${filteredSources.length}")
     filteredSources.map {
-        case SimpleSource(name, title, source, _, styles) =>
-          SimpleOgcLayer(name, title, p.crs, source, None)
-        case MapAlgebraSource(name, title, sources, algebra, _, styles) =>
-          val simpleLayers = sources.mapValues { rs => SimpleOgcLayer(name, title, p.crs, rs, None) }
-          MapAlgebraOgcLayer(name, title, p.crs, simpleLayers, algebra, None)
+        case SimpleSource(name, title, source, _, styles, resampleMethod) =>
+          SimpleOgcLayer(name, title, p.crs, source, None, resampleMethod)
+        case MapAlgebraSource(name, title, sources, algebra, _, styles, resampleMethod) =>
+          val simpleLayers = sources.mapValues { rs => SimpleOgcLayer(name, title, p.crs, rs, None, resampleMethod) }
+          MapAlgebraOgcLayer(name, title, p.crs, simpleLayers, algebra, None, resampleMethod)
       }
   }
 }